Full proper package directory structure
