--- protocol/src/main/kotlin/MessageKinds/Close.kt
@@ -1,2 +1,0 @@
-package MessageKinds
-
