--- protocol/src/main/kotlin/MessageKinds/Quote.kt
@@ -1,2 +1,0 @@
-package MessageKinds
-
