+++ protocol/src/main/kotlin/website/tbd/tbdex/protocol/Mapper.kt
@@ -1,3 +1,5 @@
+package website.tbd.tbdex.protocol
+
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.databind.ObjectReader
 import com.fasterxml.jackson.databind.ObjectWriter
