+++ protocol/src/test/kotlin/website/tbd/tbdex/protocol/TestData.kt
@@ -1,6 +1,13 @@
 package website.tbd.tbdex.protocol
 
 import typeid.TypeID
+import website.tbd.tbdex.protocol.message_kinds.Rfq
+import website.tbd.tbdex.protocol.message_kinds.RfqData
+import website.tbd.tbdex.protocol.message_kinds.SelectedPaymentMethod
+import website.tbd.tbdex.protocol.resource_kinds.CurrencyDetails
+import website.tbd.tbdex.protocol.resource_kinds.Offering
+import website.tbd.tbdex.protocol.resource_kinds.OfferingData
+import website.tbd.tbdex.protocol.resource_kinds.PresentationExchange
 
 object TestData {
   val alice = "alice"
